refactoring : move account components
