+++ backoffice/src/App.jsx
@@ -15,7 +15,7 @@ import ForgottenPassword from './components/login/ForgottenPassword';
 import LoginForm from './components/login/LoginForm';
 import LoginWithAccessToken from './components/login/LoginWithAccessToken';
 import ModerationRoutes from './components/backoffice/moderation/ModerationRoutes';
-import { MyAccount } from './components/backoffice/common/account/MyAccount';
+import { MyAccount } from './components/backoffice/account/MyAccount';
 import './utils/moment-fr';
 import './App.scss';
