+++ backoffice/src/components/backoffice/account/ChangePassword.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import './changePassword.scss';
-import { checkPasswordComplexity } from '../../../../utils/password';
-import { updatePassword } from '../../../../lib/accountService';
+import { checkPasswordComplexity } from '../../../utils/password';
+import { updatePassword } from '../../../lib/accountService';
 
 export class ChangePassword extends React.Component {
